+++ timsieved/parser.c
@@ -126,7 +126,7 @@ static struct protocol_t sieve_protocol =
       { "\"STARTTLS\"", CAPA_STARTTLS },
       { NULL, 0 } } },
   { "STARTTLS", "OK", "NO", 1 },
-  { "AUTHENTICATE", INT_MAX, 1, "OK", "NO", NULL, "*",
+  { "AUTHENTICATE", USHRT_MAX, 1, "OK", "NO", NULL, "*",
     &sieve_parsesuccess, AUTO_CAPA_AUTH_SSF },
   { NULL, NULL, NULL },
   { NULL, NULL, NULL },
